requirements.txt
asgiref==3.5.2
bidict==0.22.0
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2==2.9.5
PyJWT==2.6.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
